requirements.txt
# Requirements untuk CNN Deteksi Vannamei
# Core dependencies yang diperlukan untuk training dan deteksi

# Deep Learning Framework
torch>=2.0.0
numpy>=1.20.0

# Image Processing
pillow>=9.0.0
opencv-python>=4.5.0

# Data Visualization and Analysis
matplotlib>=3.5.0
pandas>=1.3.0

# Machine Learning
scikit-learn>=1.0.0

# System Monitoring
psutil>=5.8.0
